Build OpenType code from .fea files
Previously, the OpenType code was specified in the .sfd files.
That code is still there, and should be remove later.
Before that removal, builds should be done using master_src.sfd
(or master_src.fea) as these builds show the ligature u0B95_u0BC2,
and not the .notdef glyph. I suspect the .notdef glyph is showing
since the u0B95_u0BC2 glyph was duplicated and a duplicate was removed.
Probably when the removal happened the OpenType code to produce
that ligature pointed to the removed glyph, and as a result,
.notdef is now shown as the output of the ligature code was removed.