Adjust pulli and language specific nuktas
When the Tamil glyphs were no longer scaled,
the spacing of the nuktas and pulli was overlooked,
causing the nuktas to touch the base character.

While touching turned out to be desired for one language,
it was not desired for the other languages. So now,
most of the languages have spacing like before the
changing of the scaling of the glyphs.

To simplify using the language specific nuktas,
specific fonts were created for each language.