Use friendly working names for naming glyphs
The font sources now use glyph names such as ka, instead of u0B95.

The OpenType code in the .fea file was disabled, as the glyph names
in that file have not been converted to friendly names. This file
will be replaced with a .feax file that will be created from scratch
(and thus use friendly names from the beginning).