Rename nukta testing data
Now, test files starting with nuktas.* are mostly for testing
different codepoints for the nukta (and the dot above).
Test files starting with languages.* are mostly for testing
the different visual styles of nuktas.

But since the nukta has issues with both codepoints and languages,
the distinctions in the test data mentioned above do overlap.