Use lighter weight Latin in most places
Previously the Latin for ThiruValluvar and Vaigai came from
Gentium Book Plus, a heavier family than Gentium Plus.

Now, the bold for ThiruValluvar and Vaigai is a little lighter
than Gentium Bold, and the regular for Vaigai comes from Gentium.

The regular for ThiruValluvar could be made lighter, but not without
causing documents to change slightly. So the regular for this font is
unchanged at the moment.